backend_db/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-multipart
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
bcrypt